chore: rename spinner dir to progress
